requirements.txt
binaryornot==0.3.0
colorlog==2.6.0
cookiecutter==0.9.1
docopt==0.6.2
docx==0.2.4
funcy==1.5
html2text==2015.2.18
Jinja2==2.7.3
kitchen==1.2.1
lxml==3.4.2
MarkupSafe==0.23
mock==1.0.1
pdfminer==20140328
Pillow==2.7.0
py==1.4.26
PySide==1.2.2
python-osinfo==0.2.1
PyYAML==3.11
scandir==0.9
tox==1.9.0
Unidecode==0.4.17
virtualenv==12.0.7
